Updated to lastest driver version

+++ remove-driver.sh
@@ -1,22 +1,46 @@
 #!/bin/bash
 
 SCRIPT_NAME="remove-driver.sh"
-SCRIPT_VERSION="20210531"
+SCRIPT_VERSION="20211212"
 
 DRV_NAME="rtl88x2bu"
-DRV_VERSION="5.8.7.4"
+DRV_VERSION="5.13.1"
 OPTIONS_FILE="88x2bu.conf"
 
 DRV_DIR="$(pwd)"
+KRNL_VERSION="$(uname -r)"
 
+clear
+echo "Running ${SCRIPT_NAME} version ${SCRIPT_VERSION}"
+
+# support for NoPrompt allows non-interactive use of this script
+NO_PROMPT=0
+
+# get the options
+while [ $# -gt 0 ]
+do
+	case $1 in
+		NoPrompt)
+			NO_PROMPT=1 ;;
+		*h|*help|*)
+			echo "Syntax $0 <NoPrompt>"
+			echo "       NoPrompt - noninteractive mode"
+			echo "       -h|--help - Show help"
+			exit 1
+			;;
+	esac
+	shift
+done
+
+# check to ensure sudo was used
 if [[ $EUID -ne 0 ]]
 then
 	echo "You must run this script with superuser (root) privileges."
-	echo "Try \"sudo ./${SCRIPT_NAME}\""
+	echo "Try: \"sudo ./${SCRIPT_NAME}\""
 	exit 1
 fi
 
-echo "Starting removal."
+echo "Starting removal..."
 
 dkms remove -m ${DRV_NAME} -v ${DRV_VERSION} --all
 RESULT=$?
@@ -30,20 +54,22 @@ then
 	rm -f /etc/modprobe.d/${OPTIONS_FILE}
 	echo "Deleting source files from /usr/src/${DRV_NAME}-${DRV_VERSION}"
 	rm -rf /usr/src/${DRV_NAME}-${DRV_VERSION}
-	echo "Deleting ${DRV_DIR}"
-	rm -rf ${DRV_DIR}
 	echo "The driver was removed successfully."
+	echo "You may now delete the driver directory if desired."
 else
 	echo "An error occurred. dkms remove error = ${RESULT}"
 	echo "Please report this error."
 	exit $RESULT
 fi
 
-read -p "Are you ready to reboot now? [y/n] " -n 1 -r
-echo    # move to a new line
-if [[ $REPLY =~ ^[Yy]$ ]]
+if [ $NO_PROMPT -ne 1 ]
 then
-    reboot
+	read -p "Do you want to reboot now? (recommended) [y/N] " -n 1 -r
+	echo
+	if [[ $REPLY =~ ^[Yy]$ ]]
+	then
+		reboot
+	fi
 fi
 
 exit 0
